Implement FullAdder
I was trying to figure out a way to do the FullAdder with two HalfAdders but never actually got around to that. Figuring out that sum was just Xor on one pair and then Xor with the result of that calculation and the leftover bit went quickly. Determining how to handle the carry bit was a bit harder.